pyproject.toml
[project]
name = "schnee"
version = "0.1.0"
description = "NFC/RFID tag authentication and encryption library"
requires-python = ">=3.14, <3.15"
authors = [
    { name = "shinaji" },
]
license = "Apache-2.0"
license-files = ["LICENSE"]
classifiers = [
    "License :: OSI Approved :: Apache Software License",
]
dependencies = [
    "pycryptodome>=3.23.0,<4",
    "pydantic>=2.12.5,<3",
    "pyscard>=2.3.1,<3",
    "python-json-logger>=4.0.0,<5",
]

[dependency-groups]
dev = [
    "nox>=2025.11.12",
    "polyfactory>=3.2.0,<4",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2,<10",
    "ruff>=0.14.10,<1",
    "ty>=0.0.8,<1",
]



[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]
line-length = 88
indent-width = 4
target-version = "py314"

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "D100",
    "D104",
    "D107",
    "D204",
    "D213",
    "D400",
    "D415",
    "T201"
]
fixable = ["ALL"]
unfixable = [
    "F841",
    "T201",
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**/*.py" = ["S101"]


[tool.ruff.lint.isort]
known-third-party = []

[tool.ty.environment]
root = ["src"]

[tool.ty.src]
include = ["src/schnee", "tests"]


[tool.pytest.ini_options]
addopts = [
    "--maxfail=2",
    "-rf",
    "--import-mode=importlib"
]
testpaths = ["tests"]
pythonpath = ["src"]
